Don't execute sripts when doc is loaded from disk
This is a security feature. As we don't know whether to trust the
document author, we don't execute the scripts automatically. If the user
considers the scripts trustworthy she can allow the execution using the
new property "AllowExecution".
Newly created scripts have this flag set -- we assume the user does trust
herself ;-)
As of now this flag is reset each time the document is loaded. A
persistence mechanism in the user settings is not yet implemented.