First step towards GUI support -- Commands for object creation
There is now a workbench providing commands to create scripted objects
 - without a shape, e.g. for general purpose calculations
 - with a 3D shape, i.e. a general purpose Part object
 - with a 2D shape, i.e. a more specialited object suitable for pading
   in PartDesign

As of now, the workbench's icon is the same as the one for the 3D object.
Using the general Python icon, may be a bit too general.

The icons are remixes of existing FreeCAD icons and the Python logo,
adapted for follow FreeCADs artwork guidelines.