First shot for a rudimentary GUI
This allows to set the allow-execution flag and displays the actual
python code of the scripted object in the main view area.
All is controlled by the task panel, so the main view with the python
editor *should* be unclosable. However, it seems that Qt only allows to
enable/disable the close-button for all MDI tabs, but not individually.
-- too bad; another workaround needs to be found.

The text editor itself is also only very basic. That's on puspose.
Eventually I want to embed FreeCAD's own python editor ot some 3rd-party
component, but not reinvent yet another editor.

Still missing is the GUI to change the input and output properties of the
scripted object. As of now, this has to be done using the context-menu of
the property grid (remember to select "show all" first, otherwise "add
property" is not available).